+++ portal-service/src/com/liferay/portal/kernel/servlet/taglib/ui/JavaScriptUIItem1.java
@@ -17,27 +17,10 @@ package com.liferay.portal.kernel.servlet.taglib.ui;
 /**
  * @author Iván Zaera
  */
-public class JavascriptMenuItem extends MenuItem implements JavascriptUIItem {
+public interface JavaScriptUIItem1 extends UIActionItem {
 
-	public String getJavascript() {
-		return _javascript;
-	}
+	public String getOnClick();
 
-	@Override
-	public String getOnClick() {
-		return _onClick;
-	}
-
-	public void setJavascript(String javascript) {
-		_javascript = javascript;
-	}
-
-	@Override
-	public void setOnClick(String onClick) {
-		_onClick = onClick;
-	}
-
-	private String _javascript;
-	private String _onClick;
+	public void setOnClick(String onClick);
 
 }
